Don't require ordering between kernel and user trace.
At the moment we always run `xperf -stop -d`, the `-d` is what makes
xperf create a "merged" trace. This is somewhat slow.

Instead, we want to process the two raw ETL files individually, and skip
the xperf merging step.

We want to process the kernel trace first, and then the user trace.
The kernel trace gives us process and thread starts and ends and their names.
The user trace gives us extra markers and jit lib mapping ops.
We want to put the information from the user trace in the right place
even if we only see it after we've processed the entirety of the kernel trace.

So we now store a bunch of process / thread data with associated timestamps
so that we can find the right place once we process the user trace, even
if the pids or tids have been reused.