Fix process renaming with --reuse-threads.
In simpleperf profiles from Android, running samply import --reuse-threads
resulted in profiles with bad process names.
The processes get their names with a COMM event that's not an exec
event - on Android you rarely ever exec. All app processes are forked
from the zygote process and then renamed to the app name.
We weren't propagating the name change into the process in the case
where we have a recycler but couldn't find an existing process for
the new name.