Fix FileContentsCursor's Seek implementation.
Seeks beyond the end of the file are valid.

This fixes the handling of partially-written Jitdump files. Without this fix,
skip_next_record would return an error when skipping a partially-written record.

Specifically, this caused the assembly view in the Firefox Profiler to show
an error when double-clicking a JS JIT frame from a partially-written file.