Fix repeated external object file reads with async lookup.
When used on macOS with debug info in .o files, `symbol_map.lookup(...).await`
was re-reading the external file over and over even when it was cached inside the
symbol map.

This didn't affect samply symbolication or samply-api because samply-api doesn't
use the async `lookup` method, it uses `lookup_sync` and calls `lookup_external` manually.