Store symbol tables separately from fxprof_processed_profile::LibraryInfo.
We already have a separate Profile::set_lib_symbol_table method so just use that always.
Most consumers just specify None, and now they don't have to do that anymore.

But the main motivation for this change was performance: We were hashing the symbol
table during calls to handle_for_lib. This fixes that.