Make create_file_cleanly responsible for creating the parent directory.
This is slightly more robust against racy deletions.
Now the "do I need to create a directory" check is done in such
a way that if the answer is "no", it's sure to stay that way, because
the created file ensures it.
In the past, if the answer was "no", it could change to "yes" before
the lock file was created.