Add arm64e variant to samply-mac-preload dylib universal binary
On macOS, there are three slices for binaries: x86_64, arm64, and
arm64e. The third one is introduced[1] and it is using ARMv8.3's PAC
etc.

Because macOS samply injects dylib for the profiled process, if the
profiled process is built with arm64e, dylib needs arm64e slice as well.
Previously dylib only had x86_64 and arm64 slices.

This patch adds an instruction to add arm64e slice to this dylib as
well. This allows running samply with the process using arm64e. While
the generated code itself is not using PAC, this is fine since arm64 is
subset of arm64e as whole code is just calling and exposing C interface.

This patch does not add a built binary intentionally to allow the upstream
developers to build and add that.

[1]: https://developer.apple.com/documentation/security/preparing-your-app-to-work-with-pointer-authentication