Allow specifying SERVER_URL
This patch adds reading envvar SERVER_URL to embed it into the served
HTML & used URLs for samply output. It is possible that we would like to
launch this samply result in the testbed server and we would like to
offer profiler result remotely. So we can launch samply with
--address 0.0.0.0 and by using this SERVER_URL, we can expose samply as
a remote symbol / profile server.