Use instant marker timing if span start and end time are the same
This commit updates the way Samply adds markers to
use instant marker timing, instead of a span marker
timing, if the start time and end time of the span
are exactly the same.

Admittedly, I don't like this change, and I would like
to revert it once a different marker file format is
potentially supported.

I think that a zero-length interval should be possible
if that is actually what was measured. But as of right
now, I have no way to emit an actual instant marker
given the current file-format restrictions.