Add truncate behavior to marker file creation
This commit ensures that .truncate(true) is set when creating marker files to clarify
that existing files should be overwritten. This is unlikely to happen with the temp
directory file format containing the PID and TID, but if it were to happen, it's better
to be safe than sorry. This also fixes the clippy::suspicious_open_options warning.